login

+++ ServletDemo/src/com/itacademy/bobkevich/servlet/entity/Person.java
@@ -16,7 +16,7 @@ public class Person {
     private String login;
     private String first_name;
     private String last_name;
-    private Integer age;
+    private Long age;
     private String mail;
     private String password;
     private PersonRole personRole;
